register_sendEmail_login_withToken_logout_onlyAdminCRUD_addToCart

+++ AngLarBackend/routes/api.php
@@ -22,25 +22,54 @@ use Illuminate\Support\Facades\Route;
 
 use App\Http\Controllers\ProductController;
 Route::get('/products', [ProductController::class, 'index']);
-Route::post('/products', [ProductController::class, 'store']);
-Route::get('/products/{id}', [ProductController::class, 'show']);
-Route::put('/products/{id}', [ProductController::class, 'update']);
-Route::delete('/products/{id}', [ProductController::class, 'destroy']);
-
-use App\Http\Controllers\CustomerController;
-
-Route::post('/register', [CustomerController::class, 'register']);
-Route::post('/login', [CustomerController::class, 'login']);
-Route::get('/customers/{id}', [CustomerController::class, 'show']);
-Route::put('/customers/{id}', [CustomerController::class, 'update']);
-Route::get('/customers', [CustomerController::class, 'getAllCustomers']);
+Route::middleware('auth:sanctum','role:admin')->group(function () {
+    Route::post('/products', [ProductController::class, 'store']);
+    Route::get('/products/{id}', [ProductController::class, 'show']);
+    Route::put('/products/{id}', [ProductController::class, 'update']);
+    Route::delete('/products/{id}', [ProductController::class, 'destroy']);
+});
 
 use App\Http\Controllers\CategoryController;
-Route::get('/categories', [CategoryController::class, 'index']);
-Route::post('/categories',[CategoryController::class,'store']);
-Route::get('/categories/{id}', [CategoryController::class, 'show']);
-Route::put('/categories/{id}', [CategoryController::class, 'update']);
-Route::delete('/categories/{id}', [CategoryController::class, 'destroy']);
+Route::get('/categories', [CategoryController::class, 'index']); // mindenkinek
+Route::middleware('auth:sanctum','role:admin')->group(function () {
+    Route::get('admin/categories/{id}', [CategoryController::class, 'show']);
+    Route::post('admin/categories', [CategoryController::class, 'store']);
+    Route::put('admin/categories/{id}', [CategoryController::class, 'update']);
+    Route::delete('admin/categories/{id}', [CategoryController::class, 'destroy']);
+});
+
+use App\Http\Controllers\CustomerController;
+Route::post('/register', [CustomerController::class, 'register']);
+Route::post('/login', [CustomerController::class, 'login']);
+
+
+Route::middleware('auth:sanctum')->post('/logout', [CustomerController::class, 'logout']);
+
+
+Route::get('/customers', [CustomerController::class, 'index']);
+Route::post('/customers', [CustomerController::class, 'store']);
+Route::get('/customers/{id}', [CustomerController::class, 'show']);
+Route::put('/customers/{id}', [CustomerController::class, 'update']);
+Route::delete('/customers/{id}', [CustomerController::class, 'destroy']);
+
+use App\Http\Controllers\MailController;
+Route::get('/send-test-email', [MailController::class, 'sendTestEmail']);
+Route::get('/send-registration-confirmation', [MailController::class, 'sendRegistrationConfirmation']);
+Route::get('/send-email-verification', [MailController::class, 'sendEmailVerification']);
+Route::get('/verify-email', [CustomerController::class, 'verifyEmail']);
+
+
+use App\Http\Controllers\ForgotPasswordController;
+use App\Http\Controllers\ResetPasswordController;
+// Route::post('/password/forgot', [ForgotPasswordController::class, 'sendResetLinkEmail']);
+// Route::post('/password/reset/{token}', [CustomerController::class, 'resetPassword']);
+// // ...
+
+Route::post('/password/forgot', [ForgotPasswordController::class, 'sendResetLinkEmail']);
+Route::post('/password/reset', [CustomerController::class, 'resetPassword'])->name('password.reset');
+
+//Route::get('/password/reset/{token}', [CustomerController::class, 'showResetForm'])->name('password.reset');
+
 
 // use App\Http\Controllers\OrderController;
 
@@ -49,3 +78,18 @@ Route::delete('/categories/{id}', [CategoryController::class, 'destroy']);
     // Route::get('/orders/{id}', [OrderController::class, 'show']);
     // Route::put('/orders/{id}', [OrderController::class, 'update']);
     // Route::delete('/orders/{id}', [OrderController::class, 'destroy']);
+
+
+
+    // use App\Http\Controllers\CheckoutController;
+    // Route::post('/checkout', [CheckoutController::class, 'checkout']); 
+
+
+
+
+
+use App\Http\Controllers\OrderController;
+Route::post('/orders', [OrderController::class, 'store']);
+
+
+
